Reworked architecture of app.
